requirements.txt
numpy
matplotlib
sphinx>=1.4.3
configparser
tox
python-decouple